feat: dashboard UI

+++ api/models/logs.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Float, Integer, String
 from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
@@ -7,7 +7,7 @@ Base = declarative_base()
 class RequestLogModel(Base):
     __tablename__ = "requests"
 
-    ts = Column(Integer, primary_key=True, index=True)
+    ts = Column(Float, primary_key=True, index=True)
     client_addr = Column(String, nullable=False)
     method = Column(String, nullable=False)
     status_code = Column(Integer, nullable=False)
@@ -17,7 +17,7 @@ class RequestLogModel(Base):
 class AppLogModel(Base):
     __tablename__ = "app_logs"
 
-    ts = Column(Integer, primary_key=True, index=True)
+    ts = Column(Float, primary_key=True, index=True)
     level = Column(String, nullable=False)
     logger = Column(String, nullable=False)
     filename = Column(String, nullable=False)
